Extract graph-resolution, clients, and daemon lifecycle into focused modules
Move resolveDefaultGraphsDirs/resolveGraphsDirs/loadGraphsOrFatal from
index.ts into src/graph-resolution.ts. Extract detectClients,
clientDisplayName, allClientChoices from init.ts into src/cli/clients.ts.
Split startDaemon into writePidFile and registerShutdownHandlers for
independent testability. Remove duplicate test suites and dead re-exports.
